POC for creating a staged disk-store configiuration
A new option to the gfsh configure disk-store command provides the
capability to add a disk-store element to the cluster configuration
before any servers exist. Server joing the cluster will then inherit the
disk-store configuration and create the appropriate directory for the
oplog files.

This is useful for persisting the PDX types to a disk-store other than
DEFAULT.

Authored-by: Ken Howe <khowe@pivotal.io>